Change project structure and engine design
