+++ phoenix/Cargo.toml
@@ -1,0 +1,20 @@
+workspace = { members = ["phoenix", "phoenix_core"] }
+[package]
+name = "phoenix_engine"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+colored = "2.1.0"
+log = "0.4.22"
+fern = { version = "0.6.2", features = ["colored"] }
+gl = "0.14.0"
+slint = "1.7.2"
+glfw = "0.58.0"
+chrono = "0.4.38"
+cgmath = "0.18.0"
+
+[build-dependencies]
+slint-build = "1.7.2"
