+++ phoenix_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "phoenix_engine"
+name = "phoenix_core"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,10 +10,6 @@ colored = "2.1.0"
 log = "0.4.22"
 fern = { version = "0.6.2", features = ["colored"] }
 gl = "0.14.0"
-slint = "1.7.2"
 glfw = "0.58.0"
 chrono = "0.4.38"
 cgmath = "0.18.0"
-
-[build-dependencies]
-slint-build = "1.7.2"
