+++ phoenix_core/src/lib.rs
@@ -1,0 +1,10 @@
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::error::Error;
+
+    #[test]
+    fn engine_test() {
+        println!("Hello World!");
+    }
+}
