--- src/lib.rs
@@ -1,38 +1,0 @@
-// Prevent console window in addition to Slint window in Windows release builds when, e.g., starting the app via file manager. Ignored on other platforms.
-#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-
-pub fn add(left: usize, right: usize) -> usize {
-    left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::error::Error;
-
-    slint::include_modules!();
-
-    #[test]
-    fn engine_test() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-
-    #[test]
-
-    fn slint_test() -> Result<(), Box<dyn Error>> {
-        let ui = AppWindow::new()?;
-    
-        ui.on_request_increase_value({
-            let ui_handle = ui.as_weak();
-            move || {
-                let ui = ui_handle.unwrap();
-                ui.set_counter(ui.get_counter() + 1);
-            }
-        });
-    
-        ui.run()?;
-
-        Ok(())
-    }
-}
